STRUM-9 - created ffmpeg_server that installs itself into a docker container that has ffmpeg installed alongside
