+++ client/package.json
@@ -59,6 +59,7 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "@types/fluent-ffmpeg": "^2.1.27",
     "@types/jest": "^29.5.12",
     "@types/lodash": "^4.17.16",
     "@types/next": "^9.0.0",
