+++ ffmpeg_server/Dockerfile
@@ -1,0 +1,33 @@
+FROM node:21-alpine3.19
+
+ARG NODE_ENV=development
+ENV NODE_ENV=${NODE_ENV}
+
+RUN mkdir -p /usr/src/app
+WORKDIR /usr/src/app
+
+# Copy package files first for better caching
+COPY package.json yarn.lock ./
+RUN yarn install --frozen-lockfile && \
+    yarn cache clean
+
+# Copy source files
+COPY . .
+
+# Build TypeScript
+RUN yarn build
+
+RUN apk upgrade -U \ 
+    && apk add ca-certificates ffmpeg libva-intel-driver \
+    && rm -rf /var/cache/*
+
+EXPOSE 8080
+CMD ["node", "dist/server.js"]
+
+
+
+
+
+
+
+
