+++ ffmpeg_server/nodemon.json
@@ -1,0 +1,6 @@
+{
+  "watch": ["src"],
+  "ext": "ts",
+  "ignore": ["src/**/*.spec.ts"],
+  "exec": "tsc && docker-compose up -d --build ffmpeg-server"
+}
