+++ ffmpeg_server/package.json
@@ -1,0 +1,29 @@
+{
+  "name": "ffmpeg-server",
+  "version": "1.0.0",
+  "description": "Server to supply ffmpeg services to guitar app server api",
+  "main": "dist/server.js",
+  "license": "MIT",
+  "scripts": {
+    "start": "node dist/server.js",
+    "build": "tsc",
+    "dev": "ts-node src/server.ts",
+    "watch": "tsc -w",
+    "dev:docker": "nodemon"
+  },
+  "dependencies": {
+    "cors": "^2.8.5",
+    "dotenv": "^16.0.3",
+    "express": "^4.21.2",
+    "fluent-ffmpeg": "^2.1.3"
+  },
+  "devDependencies": {
+    "@types/cors": "^2.8.13",
+    "@types/express": "^4.17.17",
+    "@types/fluent-ffmpeg": "^2.1.27",
+    "@types/node": "^18.15.11",
+    "nodemon": "^3.1.9",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.0.3"
+  }
+}
