+++ ffmpeg_server/src/routes/index.ts
@@ -1,0 +1,7 @@
+import healthRouter from './health';
+import conversionsRouter from './conversions';
+
+export {
+    healthRouter,
+    conversionsRouter
+};
