--- server/docker/docker-compose.yaml
@@ -1,10 +1,0 @@
-version: '3.9'
-
-services:
-  ffmpg-server:
-    build:
-      context: .
-    ports:
-      - '8080:8080'
-    volumes:
-      - './data:/app/data'
