--- server/docker/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "ffmpeg-server",
-  "version": "1.0.0",
-  "description": "Server to supply ffmpeg services to guitar app server api",
-  "main": "server.js",
-  "license": "MIT",
-  "scripts": {
-    "start": "node server.js"
-  },
-  "dependencies": {
-    "cors": "^2.8.5",
-    "express": "^4.21.2",
-    "fluent-ffmpeg": "^2.1.3"
-  }
-}
