+++ server/src/routes/index.ts
@@ -1,0 +1,7 @@
+import appRouter from './app_routes';
+import facebookRouter from './facebook_routes';
+
+export {
+    appRouter,
+    facebookRouter
+};
