--- server/src/types.ts
@@ -1,25 +1,0 @@
-export interface Scale {
-  id: string;
-  name: string;
-  intervals: number[];
-}
-
-export interface ApiResponse<T> {
-  data: T;
-  error?: string;
-}
-
-export interface FacebookReel {
-  id: string;
-  title: string;
-  url: string;
-  thumbnailUrl?: string;
-  downloaded: boolean;
-  downloadPath?: string;
-  createdAt: string;
-}
-
-export interface DownloadReelRequest {
-  url: string;
-  title?: string;
-}
